refactor: consolidate contract extraction and cleanup
- Remove obsolete init-package-repos.sh script (repos already initialized)
- Remove unused extract-contracts.ts (NestJS/database-dependent version)
- Rename extract-contracts-standalone.ts to extract-contracts.ts
- Update all references in package.json and GitHub workflows
- Remove "Reactions" from planned capabilities (already implemented)
- Simplify npm scripts: extract:contracts now references unified extractor
