+++ package.json
@@ -19,8 +19,7 @@
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
     "extract:contracts": "ts-node -r tsconfig-paths/register tools/extract-contracts.ts",
-    "extract:contracts:standalone": "ts-node -r tsconfig-paths/register tools/extract-contracts-standalone.ts",
-    "generate:all": "npm run extract:contracts:standalone && npm run validate:contracts:types && npm run generate:sdk && npm run generate:cli && npm run generate:n8n && npm run generate:openapi && npm run validate:generated",
+    "generate:all": "npm run extract:contracts && npm run validate:contracts:types && npm run generate:sdk && npm run generate:cli && npm run generate:n8n && npm run generate:openapi && npm run validate:generated",
     "validate:generated": "cd generated/sdk && npm install && npm run build && npm pack && cd ../cli && npm install ../sdk/gatekit-sdk-*.tgz && npm run build && cd ../n8n && npm install && npm run build && echo '✅ Generated packages validation completed'",
     "validate:contracts:types": "ts-node -r tsconfig-paths/register tools/scripts/validate-contracts.ts",
     "generate:sdk": "ts-node -r tsconfig-paths/register tools/generators/sdk-generator.ts",
